pyproject.toml
[tool.poetry]
name = "mqtt-docker-python"
version = "0.1.0"
description = ""
authors = ["James Bristow <james.bristow@plantandfood.co.nz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
paho-mqtt = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
